pyproject.toml
[project]
name = "womp-womp"
version = "0.1.0"
description = "WHWHRHRHRRHHRRHRH KERCHUNKCHCUHNKCHUNK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blinker==1.9.0",
    "click==8.2.1",
    "flask==3.1.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "py-xid==1.0.0",
    "python-dotenv==1.1.0",
    "werkzeug==3.1.3",
]
